AFDM(FactoMineR)	R Documentation

Multiple Factor Analysis for Mixed Data

Description

Performs Multiple Factor Analysis with both quantitative and qualitative data.

Usage


AFDM (base, type, ncp = 5, graph = TRUE, sup.var = NULL, 
    ind.sup = NULL, axes = c(1,2))

Arguments

base 	 a data frame with n rows (individuals) and p columns

type 	 a vector indicating the type of each variable (the length of type must be equal to the number of columns of X

ncp 	 number of dimensions kept in the results (by default 5)

graph 	 boolean, if TRUE a graph is displayed

ind.sup 	 a vector indicating the indexes of the supplementary individuals

sup.var 	 a vector indicating the indexes of the supplementary variables

axes 	 a length 2 vector specifying the components to plot

Value

Returns a list including:

eig 	 a numeric vector containing all the eigenvalues

link.group

group

ind 	 a list of matrices with all the results for the individuals (coordinates, square cosine, contributions)

quali.var 	 a list of matrices with all the results for the qualitative variables (coordinates, square cosine, contributions, v.test)

quanti.var 	 a list of matrices with all the results for the quantitative variables (coordinates, correlation, square cosine, contributions)

call 	 a list with some statistics

Returns the individuals factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

References

Pagès J. (2004). Analyse factorielle de données mixtes. Revue Statistique Appliquée. LII (4). pp. 93-111.

See Also

print.AFDM, plot.AFDM

Examples


data(wine)
res.afdm = AFDM(wine[,c(1,2,30,31)],type=c("n","n","s","s"))

[Package FactoMineR version 1.02 Index]