FAMD {FactoMineR}	R Documentation

Factor Analysis for Mixed Data

Description

FAMD is a principal component method dedicated to explore data with both continuous and categorical variables. It can be seen roughly as a mixed between PCA and MCA. More precisely, the continuous variables are scaled to unit variance and the categorical variables are transformed into a disjunctive data table (crisp coding) and then scaled using the specific scaling of MCA. This ensures to balance the influence of both continous and categorical variables in the analysis. It means that both variables are on a equal foot to determine the dimensions of variability. This method allows one to study the similarities between individuals taking into account mixed variables and to study the relationships between all the variables. It also provides graphical outputs such as the representation of the individuals, the correlation circle for the continuous variables and representations of the categories of the categorical variables, and also specific graphs to visulaize the associations between both type of variables.

Usage

FAMD (base, ncp = 5, graph = TRUE, sup.var = NULL, 
    ind.sup = NULL, axes = c(1,2), row.w = NULL, tab.comp = NULL)

Arguments

base
a data frame with n rows (individuals) and p columns

ncp
number of dimensions kept in the results (by default 5)

graph
boolean, if TRUE a graph is displayed

ind.sup
a vector indicating the indexes of the supplementary individuals

sup.var
a vector indicating the indexes of the supplementary variables

axes
a length 2 vector specifying the components to plot

row.w
an optional row weights (by default, uniform row weights); the weights are given only for the active individuals

tab.comp
object obtained from the imputeFAMD function of the missMDA package that allows to handle missing values

Value

Returns a list including:

eig
a matrix containing all the eigenvalues, the percentage of variance and the cumulative percentage of variance

var
a list of matrices containing all the results for the variables considered as group (coordinates, square cosine, contributions)

ind
a list of matrices with all the results for the individuals (coordinates, square cosine, contributions)

quali.var
a list of matrices with all the results for the categorical variables (coordinates, square cosine, contributions, v.test)

quanti.var
a list of matrices with all the results for the quantitative variables (coordinates, correlation, square cosine, contributions)

call
a list with some statistics

Returns the individuals factor map.

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr

References

Pages J. (2004). Analyse factorielle de donnees mixtes. Revue Statistique Appliquee. LII (4). pp. 93-111.

See Also

print.FAMD, summary.FAMD, plot.FAMD,
Video showing how to perform FAMD with FactoMineR

Examples

## Not run: 
data(geomorphology)
res <- FAMD(geomorphology)
summary(res)

data(wine)
res <- FAMD(wine[,c(1,2,30,31)])

## End(Not run)

[Package FactoMineR version 1.41 Index]