RegBest {FactoMineR}	R Documentation

Select variables in multiple linear regression

Description

Find an optimal submodel

Usage

RegBest(y,x, int = TRUE, wt=NULL, na.action = na.omit,
    method=c("r2","Cp", "adjr2"), nbest=1)

Arguments

y
A response vector

x
A matrix of predictors

int
Add an intercept to the model

wt
Optional weight vector

na.action
Handling missing values

method
Calculate R-squared, adjusted R-squared or Cp to select the model. By default a the F-test on the r-square is used

nbest
number of best models for each set of explained variables (by default 1)

Value

Returns the objects

all
gives all the nbest best models for a given number of variables

best
the best model

Author(s)

Francois Husson husson@agrocampus-ouest.fr

See Also

lm

Examples

data(milk)
res = RegBest(y=milk[,6],x=milk[,-6])
res$best

[Package FactoMineR version 1.41 Index]