autoLab {FactoMineR}	R Documentation

Function to better position the labels on the graphs

Description

Function to better position the labels on the graphs.

Usage

autoLab(x, y = NULL, labels = seq(along = x), cex = 1,
                       method = c("SANN", "GA"),
                       allowSmallOverlap = FALSE,
                       trace = FALSE, shadotext = FALSE,
                       doPlot = TRUE, ...)

Arguments

x
the x-coordinates

y
the y-coordinates

labels
the labels

cex
cex

method
not used

allowSmallOverlap
boolean

trace
boolean

shadotext
boolean

doPlot
boolean

...
further arguments passed to or from other methods

Value

See the text function

[Package FactoMineR version 1.41 Index]