catdes {FactoMineR}	R Documentation

Categories description

Description

Description of the categories of one factor by categorical variables and/or by quantitative variables

Usage

catdes(donnee,num.var,proba = 0.05, row.w = NULL, na.method="NA")

Arguments

donnee
a data frame made up of at least one categorical variables and a set of quantitative variables and/or categorical variables

num.var
the indice of the variable to characterized

proba
the significance threshold considered to characterized the category (by default 0.05)

row.w
a vector of integers corresponding to an optional row weights (by default, a vector of 1 for uniform row weights)

na.method
a boolean that says how to manage missing values. If ; if na.method="NA" a new category is considered for the categorical variable; if na.method="na.omit" the missing values are deleted

Value

Returns a list including:

test.chi
The categorical variables which characterized the factor are listed in ascending order (from the one which characterized the most the factor to the one which significantly characterized with the proba proba

category
description of each category of the num.var by each category of all the categorical variables

quanti.var
the global description of the num.var variable by the quantitative variables with the square correlation coefficient and the p-value of the F-test in a one-way analysis of variance (assuming the hypothesis of homoscedsticity)

quanti
the description of each category of the num.var variable by the quantitative variables.

Author(s)

Francois Husson francois.husson@institut-agro.fr

References

Husson, F., Le, S. and Pages, J. (2010). Exploratory Multivariate Analysis by Example Using R, Chapman and Hall. Lebart, L., Morineau, A. and Piron, M. (1995) Statistique exploratoire multidimensionnelle, Dunod.

See Also

plot.catdes, condes

Examples

data(wine)
catdes(wine, num.var=2)

[Package FactoMineR version 2.13 Index]