coeffRV {FactoMineR}	R Documentation

Calculate the RV coefficient and test its significance

Description

Calculate the RV coefficient and test its significance.

Usage

coeffRV(X, Y)

Arguments

X
a matrix with n rows (individuals) and p numerous columns (variables)

Y
a matrix with n rows (individuals) and p numerous columns (variables)

Details

Calculates the RV coefficient between X and Y. It returns also the standardized RV, the expectation, the variance and the skewness under the permutation distribution. These moments are used to approximate the exact distribution of the RV statistic with the Pearson type III approximation and the p-value associated to this test is given.

Value

A list containing the following components:

RV
the RV coefficient between the two matrices

RVs
the standardized RV coefficients

mean
the mean of the RV permutation distribution

variance
the variance of the RV permutation distribution

skewness
the skewness of the RV permutation distribution

p.value
the p-value associated to the test of the significativity of the RV coefficient (with the Pearson type III approximation

Author(s)

Julie Josse, Francois Husson Francois.Husson@agrocampus-ouest.fr

References

Escouffier, Y. (1973) Le traitement des variables vectorielles. Biometrics 29 751–760.
Josse, J., Husson, F., Pag\'es, J. (2007) Testing the significance of the RV coefficient. Computational Statististics and Data Analysis. 53 82–91.
Kazi-Aoual, F., Hitier, S., Sabatier, R., Lebreton, J.-D., (1995) Refined approximations to permutations tests for multivariate inference. Computational Statistics and Data Analysis, 20, 643–656

Examples

data(wine)
X <- wine[,3:7]
Y <- wine[,11:20]
coeffRV(X,Y)

[Package FactoMineR version 1.41 Index]