condes {FactoMineR}	R Documentation

Continuous variable description

Description

Description continuous by quantitative variables and/or by categorical variables

Usage

condes(donnee,num.var,weights=NULL,proba = 0.05)

Arguments

donnee
a data frame made up of at least one quantitative variable and a set of quantitative variables and/or categorical variables

num.var
the number of the variable to characterized

weights
weights for the individuals; if NULL, all individuals has a weight equals to 1; the sum of the weights can be equal to 1 and then the weights will be multiplied by the number of individuals, the sum can be greater than the number of individuals

proba
the significance threshold considered to characterized the category (by default 0.05)

Value

Returns a list including:

quanti
the description of the num.var variable by the quantitative variables. The variables are sorted in ascending order (from the one which characterized the most to the one which significantly characterized with the proba proba)

quali
The categorical variables which characterized the continuous variables are listed in ascending order

category
description of the continuous variable num.var by each category of all the categorical variables

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr

See Also

catdes

Examples

data(decathlon)
condes(decathlon, num.var=3)

[Package FactoMineR version 1.41 Index]