descfreq {FactoMineR}	R Documentation

Description of frequencies

Description

Description of the rows of a contingency table or of groups of rows of a contingency table

Usage

descfreq(donnee, by.quali = NULL, proba = 0.05)

Arguments

donnee
a data frame corresponding to a contingency table (quantitative data)

by.quali
a factor used to merge the data from different rows of the contingency table; by default NULL and each row is characterized

proba
the significance threshold considered to characterized the category (by default 0.05)

Value

Returns a list with the characterization of each rows or each group of the by.quali. A test corresponding to the hypergeometric distribution is performed and the probability to observe a more extreme value than the one observed is calculated. For each row (or category), each of the columns characterising the row are sorted in ascending order of p-value.

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr

References

Lebart, L., Morineau, A. and Piron, M. (1995) Statistique exploratoire multidimensionnelle, Dunod.

See Also

catdes, condes, textual

Examples

data(children)
descfreq(children[1:14,1:5])    ## desc of rows
descfreq(t(children[1:14,1:5])) ## desc of columns

[Package FactoMineR version 1.41 Index]