dimdesc {FactoMineR}	R Documentation

Dimension description

Description

This function is designed to point out the variables and the categories that are the most characteristic according to each dimension obtained by a Factor Analysis.

Usage

dimdesc(res, axes = 1:3, proba = 0.05)

Arguments

res
an object of class PCA, MCA, CA, MFA or HMFA

axes
a vector with the dimensions to describe

proba
the significance threshold considered to characterized the dimension (by default 0.05)

Value

Returns a list including:

quanti
the description of the dimensions by the quantitative variables. The variables are sorted.

quali
the description of the dimensions by the categorical variables

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr

References

Husson, F., Le, S. and Pages, J. (2010). Exploratory Multivariate Analysis by Example Using R, Chapman and Hall.

See Also

PCA, CA, MCA, MFA, HMFA,
Video showing how to use this function

Examples

data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup=13, graph=FALSE)
dimdesc(res.pca)

[Package FactoMineR version 1.41 Index]