estim_ncp {FactoMineR}	R Documentation

Estimate the number of components in Principal Component Analysis

Description

Estimate the number of components in PCA .

Usage

estim_ncp(X, ncp.min=0, ncp.max=NULL, scale=TRUE, method="GCV")

Arguments

X
a data frame with continuous variables

ncp.min
minimum number of dimensions to interpret, by default 0

ncp.max
maximum number of dimensions to interpret, by default NULL which corresponds to the number of columns minus 2

scale
a boolean, if TRUE (value set by default) then data are scaled to unit variance

method
method used to estimate the number of components, "GCV" for the generalized cross-validation approximation or "Smooth" for the smoothing method (by default "GCV")

Value

Returns ncp the best number of dimensions to use (find the minimum or the first local minimum) and the mean error for each dimension tested

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr, Julie JosseJulie.Josse@agrocampus-ouest.fr

References

Josse, J. and Husson, F. (2012). Selecting the number of components in PCA using cross-validation approximations. Computational Statistics and Data Analysis, 56, 1869-1879.

See Also

PCA

Examples

data(decathlon)
nb.dim <- estim_ncp(decathlon[,1:10],scale=TRUE)

[Package FactoMineR version 1.41 Index]