GPA {FactoMineR}	R Documentation

Generalised Procrustes Analysis

Description

Performs Generalised Procrustes Analysis (GPA) that takes into account missing values.

Usage

GPA(df, tolerance=10^-10, nbiteration=200, scale=TRUE, 
    group, name.group = NULL, graph = TRUE, axes = c(1,2))

Arguments

df
a data frame with n rows (individuals) and p columns (quantitative varaibles)

tolerance
a threshold with respect to which the algorithm stops, i.e. when the difference between the GPA loss function at step n and n+1 is less than tolerance

nbiteration
the maximum number of iterations until the algorithm stops

scale
a boolean, if TRUE (which is the default value) scaling is required

group
a vector indicating the number of variables in each group

name.group
a vector indicating the name of the groups (the groups are successively named group.1, group.2 and so on, by default)

graph
boolean, if TRUE a graph is displayed

axes
a length 2 vector specifying the components to plot

Details

Performs a Generalised Procrustes Analysis (GPA) that takes into account missing values: some data frames of df may have non described or non evaluated rows, i.e. rows with missing values only.
The algorithm used here is the one developed by Commandeur.

Value

A list containing the following components:

RV
a matrix of RV coefficients between partial configurations

RVs
a matrix of standardized RV coefficients between partial configurations

simi
a matrix of Procrustes similarity indexes between partial configurations

scaling
a vector of isotropic scaling factors

dep
an array of initial partial configurations

consensus
a matrix of consensus configuration

Xfin
an array of partial configurations after transformations

correlations
correlation matrix between initial partial configurations and consensus dimensions

PANOVA
a list of "Procrustes Analysis of Variance" tables, per assesor (config), per product(objet), per dimension (dimension)

Author(s)

Elisabeth Morand

References

Commandeur, J.J.F (1991) Matching configurations.DSWO press, Leiden University.
Dijksterhuis, G. & Punter, P. (1990) Interpreting generalized procrustes analysis "Analysis of Variance" tables, Food Quality and Preference, 2, 255–265
Gower, J.C (1975) Generalized Procrustes analysis, Psychometrika, 40, 33–50
Kazi-Aoual, F., Hitier, S., Sabatier, R., Lebreton, J.-D., (1995) Refined approximations to permutations tests for multivariate inference. Computational Statistics and Data Analysis, 20, 643–656
Qannari, E.M., MacFie, H.J.H, Courcoux, P. (1999) Performance indices and isotropic scaling factors in sensory profiling, Food Quality and Preference, 10, 17–21

Examples

## Not run: 
data(wine)
res.gpa <- GPA(wine[,-(1:2)], group=c(5,3,10,9,2),
    name.group=c("olf","vis","olfag","gust","ens"))

### If you want to construct the partial points for some individuals only
plotGPApartial (res.gpa)

## End(Not run)

[Package FactoMineR version 1.41 Index]