graph.var {FactoMineR}	R Documentation

Make graph of variables

Description

Plot the graphs of the variables after a Factor Analysis.

Usage

graph.var(x, axes = c(1, 2), 
    xlim = NULL, ylim = NULL, col.sup = "blue", 
    col.var = "black", draw="all", label=draw, lim.cos2.var = 0.1,
    cex = 1, title = NULL, new.plot = TRUE, ...)
    

Arguments

x
an object of class PCA, MCA, MFA or HMFA

axes
a length 2 vector specifying the components to plot

xlim
range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim
range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

col.sup
a color for the quantitative supplementary variables

col.var
a color for the variables

draw
a list of character for the variables which are drawn (by default, all the variables are drawn). You can draw all the active variables by putting "var" and/or all the supplementary variables by putting "quanti.sup" and/or a list with the names of the variables which should be drawn

label
a list of character for the variables which are labelled (by default, all the drawn variables are labelled). You can label all the active variables by putting "var" and/or all the supplementary variables by putting "quanti.sup" and/or a list with the names of the variables which should be labelled

lim.cos2.var
value of the square cosinus under the variables are not drawn

cex
cf. function par in the graphics package

title
string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

new.plot
boolean, if TRUE, a new graphical device is created

...
further arguments passed to or from other methods

Value

Returns the variables factor map.

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr

See Also

PCA, MFA, MCA, DMFA, HMFA

Examples

data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup = 13, graph = FALSE)
graph.var (res.pca, draw = c("var","Points"), 
    label = c("Long.jump", "Points"))

[Package FactoMineR version 1.41 Index]