meansComp {FactoMineR}	R Documentation

Perform pairwise means comparisons

Description

Perform means comparisons and give groups of means that are not significantly different.

Usage

meansComp(object, spec, graph=TRUE, ...) 
    

Arguments

object
A fitted model object that is supported, such as the result of a call to LinearModel, lm or aov.

spec
spec may also be a formula or a list (optionally named) of valid specs. Use of formulas is described in the Overview section below.

graph
Boolean; plot the graph to compare the means.

...
other arguments, cf the function emmeans.

Author(s)

Francois Husson francois.husson@institut-agro.fr

See Also

plot.meansComp

Examples

  data(senso)
  res <- LinearModel(Score~ Product + Day , data=senso, selection="none")
  meansComp(res,~Product)
  
## Not run: 
  ## and with the sidak correction
  meansComp(res,~Product,adjust="sidak")

## End(Not run)

[Package FactoMineR version 2.13 Index]