plot.AFDM(FactoMineR)	R Documentation

Draw the Multiple Factor Analysis for Mixt Data graphs

Description

Draw the Multiple Factor Analysis for Mixt Data graphs.

Usage


## S3 method for class 'AFDM':
plot(x, choix = "group", axes = c(1, 2), lab.grpe = TRUE,
    lab.var = TRUE, lab.ind = TRUE, habillage = "none", col.lab = FALSE,
    col.hab = NULL, invisible = NULL, lim.cos2.var = 0.1, xlim = NULL,
    ylim = NULL, cex = 1, title = NULL, ...)

Arguments

x 	 an object of class AFDM

choix 	 a string corresponding to the graph that you want to do ("ind" for the individual or qualitative variables graph, "group" for all the variables (quantitative and qualitative), "var" for the correlation circle)

axes 	 a length 2 vector specifying the components to plot

lab.grpe 	 boolean, if TRUE, the label of the groups are drwan

lab.var 	 boolean indicating if the labelled of the variables should be drawn on the map

lab.ind 	 boolean indicating if the labelled of the individuals should be drawn on the map

habillage 	 string corresponding to the color which are used. If "ind", one color is used for each individual else if it is the name or the position of a qualitative variable, it colors according to the different categories of this variable

col.lab 	 boolean indicating if the labelled should be colored

col.hab 	 vector indicating the colors to use to labelled the rows or columns elements chosen in habillage

invisible 	 list of string; for choix ="ind", the individuals can be omit (invisible = "ind"), or supplementary individuals (invisible="ind.sup") or the centerg of gravity of the qualitative variables (invisible= "quali"); if invisible = c("ind","ind.sup"), just the centers of gravity are drawn

lim.cos2.var 	 value of the square cosinus under the variables are not drawn

xlim 	 range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim 	 range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

cex 	 cf. function par in the graphics package

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

... 	 further arguments passed to or from other methods

Value

Returns the individuals factor map and the variables factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

AFDM

[Package FactoMineR version 1.04 Index]