plot.GPA {FactoMineR}	R Documentation

Draw the General Procrustes Analysis (GPA) map

Description

Draw the General Procrustes Analysis (GPA) map.

Usage

## S3 method for class 'GPA'
plot(x, axes = c(1, 2), 
    lab.ind.moy = TRUE, habillage = "ind",
    partial = "all", chrono = FALSE, xlim = NULL, ylim = NULL, 
    cex = 1, title = NULL, palette = NULL, ...)

Arguments

x
an object of class GPA

axes
a length 2 vector specifying the components to plot

lab.ind.moy
boolean, if TRUE, the label of the mean points are drawn

habillage
string corresponding to the color which are used. If "ind", one color is used for each individual; if "group" the individuals are colored according to the group

partial
list of the individuals or of the center of gravity for which the partial points should be drawn (by default, partial = "none" and no partial points are drawn)

chrono
boolean, if TRUE, the partial points of a same point are linked (useful when groups correspond to different moment)

xlim
range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim
range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

cex
cf. function par in the graphics package

title
string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

palette
the color palette used to draw the points. By default colors are chosen. If you want to define the colors : palette=palette(c("black","red","blue")); or you can use: palette=palette(rainbow(30)), or in black and white for example: palette=palette(gray(seq(0,.9,len=25)))

...
further arguments passed to or from other methods

Value

Returns the General Procrustes Analysis map.

Author(s)

Elisabeth Morand, Francois Husson Francois.Husson@agrocampus-ouest.fr

See Also

GPA

[Package FactoMineR version 1.41 Index]