predict.LinearModel {FactoMineR}	R Documentation

Predict method for Linear Model Fits

Description

Predicted values based on LinearModel object.

Usage

## S3 method for class 'LinearModel'
predict(object, newdata, interval = c("none", "confidence", "prediction"),
        level = 0.95, type = c("response", "terms"), ...)

Arguments

object
Object of class inheriting from "LinearModel"

newdata
An optional data frame in which to look for variables with which to predict. If omitted, the fitted values are used.

interval
Type of interval calculation. Can be abbreviated.

level
Tolerance/confidence level.

type
Type of prediction (response or model term). Can be abbreviated.

...
further arguments passed to or from other methods such as lm.

Details

See the help of predict.lm function.

Author(s)

Francois Husson francois.husson@institut-agro.fr

See Also

LinearModel

[Package FactoMineR version 2.13 Index]