predict.MCA {FactoMineR}	R Documentation

Predict projection for new rows with Multiple Correspondence Analysis

Description

Predict the projection of new rows with Multiple Correspondence Analysis.

Usage

## S3 method for class 'MCA'
predict(object, newdata, ...)

Arguments

object
an object of class MCA

newdata
A data frame or a matrix in which to look for variables with which to predict. newdata must contain columns with the same names as the original data.

...
Other options.

Author(s)

Francois Husson husson@agrocampus-ouest.fr

See Also

MCA

[Package FactoMineR version 1.41 Index]