print.CA(FactoMineR)	R Documentation

Print the Correspondance Analysis (CA) results

Description

Print the Correspondance Analysis (CA) results.

Usage


## S3 method for class 'CA':
print(x, file = NULL, sep = ";", ...)

Arguments

x 	 an object of class CA

file 	 A connection, or a character string naming the file to print to. If NULL (the default), the results are not printed in a file

sep 	 character string to insert between the objects to print (if the argument file is not NULL

... 	 further arguments passed to or from other methods

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

CA, write.infile

[Package FactoMineR version 1.00 Index]