print.CaGalt {FactoMineR}	R Documentation

Print the Correspondence Analysis on Generalised Aggregated Lexical Table (CaGalt) results

Description

Print the Correspondence Analysis on Generalised Aggregated Lexical Table (CaGalt) results

Usage

## S3 method for class 'CaGalt'
## S3 method for class 'CaGalt'
print(x, file = NULL, sep = ";", ...)

Arguments

x
an object of class CaGalt

file
A connection, or a character string naming the file to print to. If NULL (the default), the results are not printed in a file

sep
character string to insert between the objects to print (if the argument file is not NULL)

...
further arguments passed to or from other methods

Author(s)

Belchin Kostov badriyan@clinic.ub.es, Monica Becue-Bertaut, Francois Husson

See Also

CaGalt, write.infile

Examples

## Not run: 
data(health)
res.cagalt<-CaGalt(Y=health[,1:115],X=health[,116:118],type="n")
print(res.cagalt)

## End(Not run)

[Package FactoMineR version 1.41 Index]