print.HCPC {FactoMineR}	R Documentation

Print the Hierarchical Clustering on Principal Components (HCPC) results

Description

Print the Hierarchical Clustering on Principal Components (HCPC) results.

Usage

## S3 method for class 'HCPC'
print(x, file = NULL, sep = ";", ...)

Arguments

x
an object of class HCPC

file
A connection, or a character string naming the file to print to. If NULL (the default), the results are not printed in a file

sep
character string to insert between the objects to print (if the argument file is not NULL

...
further arguments passed to or from other methods

Author(s)

Francois Husson husson@agrocampus-ouest.fr

See Also

HCPC, write.infile

[Package FactoMineR version 1.41 Index]