print.LinearModel {FactoMineR}	R Documentation

Print the LinearModel results

Description

Print the results of the ANOVA obtained by the function LinearModel.

Usage

## S3 method for class 'LinearModel'
print(x, ...)

Arguments

x
an object of class LinearModel

...
further arguments passed to or from other methods

Details

Gives the results of the LinearModel function. If a model selection is performed, the global F-test for the complete model is first given, then all the results are given for the selected model (global F-test, the F-tests for main effects and interaction, the t-tests)

Author(s)

Francois Husson francois.husson@institut-agro.fr

See Also

LinearModel

Examples

## Not run: 
data(senso)
res <- LinearModel(Score~ Product + Day , data=senso)
res

res2 <- LinearModel(Score~ Product + Day , data=senso, selection="BIC")
res2

## End(Not run)

[Package FactoMineR version 2.13 Index]