reconst {FactoMineR}	R Documentation

Reconstruction of the data from the PCA, CA or MFA results

Description

Reconstruct the data from the PCA, CA or MFA results.

Usage

reconst(res, ncp=NULL)

Arguments

res
an object of class PCA, CA or MFA

ncp
number of dimensions used to reconstitute the data (by default NULL and the number of dimensions calculated for the PCA, CA or MFA is used)

Value

Returns a data frame with the number of individuals and the number of variables used for the PCA, CA or MFA

Author(s)

Francois Husson Francois.Husson@agrocampus-ouest.fr, Julie JosseJulie.Josse@agrocampus-ouest.fr

See Also

PCA,CA, MFA

Examples

data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup=13, graph=FALSE)
rec <- reconst(res.pca,ncp=2)

[Package FactoMineR version 1.41 Index]