summary.CA {FactoMineR}	R Documentation

Printing summeries of ca objects

Description

Printing summaries of correspondence analysis objects

Usage

## S3 method for class 'CA'
summary(object, nb.dec = 3, nbelements=10,
   ncp = 3, align.names=TRUE, file="", ...)

Arguments

object
an object of class CA

nb.dec
number of decimal printed

nbelements
number of elements written (rows, columns, ...) ; use nbelements = Inf if you want to have all the elements

ncp
number of dimensions printed

align.names
boolean, if TRUE the names of the objects are written using the same number of characters

file
a connection, or a character string naming the file to print to

...
further arguments passed to or from other methods, such as cex, cex.main, ...

Author(s)

Francois Husson husson@agrocampus-ouest.fr

See Also

CA

[Package FactoMineR version 1.41 Index]