summary.CaGalt {FactoMineR}	R Documentation

Printing summaries of CaGalt objects

Description

Printing summaries of Correspondence Analysis on Generalised Aggregated Lexical Table (CaGalt) objects

Usage

## S3 method for class 'CaGalt'
summary(object, nb.dec = 3, nbelements=10, nbind = nbelements, 
  ncp = 3, align.names=TRUE, file="", ...)

Arguments

object
an object of class CaGalt

nb.dec
number of printed decimals

nbelements
number of written elements (variables, categories, frequencies); use nbelements = Inf if you want to have all the elements

nbind
number of written elements; use nbind = Inf to have the results for all the individuals and nbind = 0 if you do not want the results for individuals

ncp
number of printed dimensions

align.names
boolean, if TRUE the names of the objects are written using the same number of characters

file
a connection, or a character string naming the file to print to

...
further arguments passed to or from other methods, such as cex, cex.main, ...

Author(s)

Belchin Kostov badriyan@clinic.ub.es, Monica Becue-Bertaut, Francois Husson

See Also

CaGalt

Examples

## Not run: 
data(health)
res.cagalt<-CaGalt(Y=health[,1:115],X=health[,116:118],type="n")
summary(res.cagalt)

## End(Not run)

[Package FactoMineR version 1.41 Index]