svd.triplet {FactoMineR}	R Documentation

Singular Value Decomposition of a Matrix

Description

Compute the singular-value decomposition of a rectangular matrix with weights for rows and columns.

Usage

svd.triplet(X, row.w=NULL, col.w=NULL, ncp=Inf)

Arguments

X
a data matrix

row.w
vector with the weights of each row (NULL by default and the weights are uniform)

col.w
vector with the weights of each column (NULL by default and the weights are uniform)

ncp
the number of components kept for the outputs

Value

vs
a vector containing the singular values of 'x';

u
a matrix whose columns contain the left singular vectors of 'x';

v
a matrix whose columns contain the right singular vectors of 'x'.

See Also

svd

[Package FactoMineR version 1.41 Index]