tab.disjonctif.prop {FactoMineR}	R Documentation

Make a disjunctive table when missing values are present

Description

Create a disjunctive table. The missing values are replaced by the proportion of the category.

Usage

tab.disjonctif.prop(tab,seed=NULL,row.w=NULL)

Arguments

tab
a data frame with factors

row.w
an optional row weights (by default, a vector of 1 for uniform row weights)

seed
a single value, interpreted as an integer for the set.seed function (if seed = NULL, missing values are initially imputed by the mean of each variable)

Value

The disjonctif table.prop

[Package FactoMineR version 1.41 Index]