CA(FactoMineR)	R Documentation

Correspondence Analysis (CA)

Description

Performs Correspondence Analysis (CA) including supplementary row and/or column points.

Usage


CA(X, ncp = 5, row.sup = NULL, col.sup = NULL, graph = TRUE,
    axes = c(1,2), row.w = NULL)

Arguments

X 	 a data frame with n rows and p columns, i.e. a contingency table

ncp 	 number of dimensions kept in the results (by default 5)

row.sup 	 a vector indicating the indexes of the supplementary rows

col.sup 	 a vector indicating the indexes of the supplementary columns

graph 	 boolean, if TRUE a graph is displayed

axes 	 a length 2 vector specifying the components to plot

row.w 	 an optional row weights (by default, uniform row weights)

Value

Returns a list including:

eig 	 a numeric vector containing all the eigenvalues

col 	 a list of matrices with all the results for the column variable (coordinates, square cosine, contributions)

row 	 a list of matrices with all the results for the row variable (coordinates, square cosine, contributions)

col.sup 	 a list of matrices containing all the results for the supplementary column points (coordinates, square cosine)

row.sup 	 a list of matrices containing all the results for the supplementary row points (coordinates, square cosine)

call 	 a list with some statistics

Returns the row and column points factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

References

BENZECRI, J.-P. (1992) Correspondence Analysis Handbook, New-York : Dekker
BENZECRI, J.-P. (1980) L'analyse des données tome 2 : l'analyse des correspondances, Paris : Bordas
GREENACRE, M.J. (1993) Correspondence Analysis in Practice, London : Academic Press

See Also

print.CA, plot.CA, dimdesc

Examples


data(children)
res.ca <- CA (children, col.sup = 6:8, row.sup = 15:18)

[Package FactoMineR version 1.04 Index]