DMFA(FactoMineR)	R Documentation

Duale Multiple Factor Analysis (DMFA)

Description

Performs Duale Multiple Factor Analysis (DMFA) with supplementary individuals, supplementary quantitative variables and supplementary qualitative variables.

Usage


DMFA(don, num.fact = ncol(data), scale.unit = TRUE, ncp = 5, 
    quanti.sup = NULL, quali.sup = NULL, graph = TRUE, axes=c(1,2))

Arguments

don 	 a data frame with n rows (individuals) and p columns (numeric variables)

num.fact 	 the number of the qualitative variable which allows to make the group of individuals

scale.unit 	 a boolean, if TRUE (value set by default) then data are scaled to unit variance

ncp 	 number of dimensions kept in the results (by default 5)

quanti.sup 	 a vector indicating the indexes of the quantitative supplementary variables

quali.sup 	 a vector indicating the indexes of the qualitative supplementary variables

graph 	 boolean, if TRUE a graph is displayed

axes 	 a length 2 vector specifying the components to plot

Value

Returns a list including:

eig 	 a numeric vector containing all the eigenvalues

var 	 a list of matrices containing all the results for the active variables (coordinates, correlation between variables and axes, square cosine, contributions)

ind 	 a list of matrices containing all the results for the active individuals (coordinates, square cosine, contributions)

ind.sup 	 a list of matrices containing all the results for the supplementary individuals (coordinates, square cosine)

quanti.sup 	 a list of matrices containing all the results for the supplementary quantitative variables (coordinates, correlation between variables and axes)

quali.sup 	 a list of matrices containing all the results for the supplementary qualitative variables (coordinates of each categories of each variables, and v.test which is a criterion with a Normal distribution)

Returns the individuals factor map and the variables factor map.

Author(s)

Sébastien Lê, François Husson François.Husson@agrocampus-rennes.fr

See Also

plot.DMFA, dimdesc

Examples


## Example with the famous Fisher's iris data
res.dmfa = DMFA ( iris, num.fact = 5)

[Package FactoMineR version 1.03 Index]