FactoMineR-package {FactoMineR}	R Documentation

Multivariate Exploratory Data Analysis and Data Mining with R

Description

The method proposed in this package are exploratory mutlivariate methods such as principal component analysis, correspondence analysis or clustering.

Details

Package: 	 FactoMineR

Type: 	 Package

Version: 	 1.34

Date: 	 2014-09-26

License: 	 GPL

LazyLoad: 	 yes

Author(s)

Francois Husson, Julie Josse, Sebastien Le, Jeremy Mazet

Maintainer: husson@agrocampus-ouest.fr

References

Le, S., Josse, J. & Husson, F. (2008). FactoMineR: An R Package for Multivariate Analysis. Journal of Statistical Software. 25(1). pp. 1-18. http://www.jstatsoft.org/v25/i01/

A website: http://factominer.free.fr/

Some videos: https://www.youtube.com/playlist?list=PLnZgp6epRBbTsZEFXi_p6W48HhNyqwxIu

[Package FactoMineR version 1.41 Index]