HMFA(FactoMineR)	R Documentation

Hierarchical Multiple Factor Analysis

Description

Performs a hierarchical multiple factor analysis, using an object of class list of data.frame.

Usage


HMFA(X,H,type = rep("s", length(H[[1]])), ncp = 5, graph = TRUE,
    axes = c(1,2), name.group = NULL)

Arguments

X 	 a data.frame

H 	 a list with one vector for each hierarchical level; in each vector the number of variables or the number of group constituting the group

type 	 the type of variables in each group in the first partition; three possibilities: "c" or "s" for quantitative variables (the difference is that for "s", the variables are scaled in the program), "n" for qualitative variables; by default, all the variables are quantitative and the variables are scaled unit

ncp 	 number of dimensions kept in the results (by default 5)

graph 	 boolean, if TRUE a graph is displayed

axes 	 a length 2 vector specifying the components to plot

name.group 	 a list of vector containing the name of the groups for each level of the hierarchy (by default, NULL and the group are named L1.G1, L1.G2 and so on)

Value

Returns a list including:

eig 	 a numeric vector with the all eigenvalues

group 	 a list of matrices with all the results for the groups (Lg and RV coefficients, coordinates, square cosine, contributions, distance to the origin, the correlations between each group and each factor)

ind 	 a list of matrices with all the results for the active individuals (coordinates, square cosine, contributions)

quanti.var 	 a list of matrices with all the results for the quantitative variables (coordinates, correlation between variables and axes)

quali.var 	 a list of matrices with all the results for the supplementary qualitative variables (coordinates of each categories of each variables, and v.test which is a criterion with a Normal distribution)

partial 	 a list of arrays with the coordinates of the partial points for each partition

Author(s)

Sébastien Lê, François Husson

References

Le Dien, S. & Pagès, J. (2003) Hierarchical Multiple factor analysis: application to the comparison of sensory profiles, Food Quality and Preferences, 18 (6), 453-464.

See Also

print.HMFA, plot.HMFA, dimdesc

Examples

 
data(wine)
hierar <- list(c(2,5,3,10,9,2), c(4,2))
res.hmfa <- HMFA(wine, H = hierar, type=c("n",rep("s",5)))

[Package FactoMineR version 1.04 Index]