MCA(FactoMineR)	R Documentation

Multiple Correspondence Analysis (MCA)

Description

Performs Multiple Correspondence Analysis (MCA) with supplementary individuals, supplementary quantitative variables and supplementary qualitative variables.

Usage


MCA(X, ncp = 5, ind.sup = NULL, quanti.sup = NULL, 
    quali.sup = NULL, graph = TRUE, axes = c(1,2),
    row.w = NULL)

Arguments

X 	 a data frame with n rows (individuals) and p columns (categorical variables)

ncp 	 number of dimensions kept in the results (by default 5)

ind.sup 	 a vector indicating the indexes of the supplementary individuals

quanti.sup 	 a vector indicating the indexes of the quantitative supplementary variables

quali.sup 	 a vector indicating the indexes of the qualitative supplementary variables

graph 	 boolean, if TRUE a graph is displayed

axes 	 a length 2 vector specifying the components to plot

row.w 	 an optional row weights (by default, uniform row weights)

Value

Returns a list including:

eig 	 a numeric vector containing all the eigenvalues

var 	 a list of matrices containing all the results for the active variables (coordinates, square cosine, contributions, v.test)

ind 	 a list of matrices containing all the results for the active individuals (coordinates, square cosine, contributions)

ind.sup 	 a list of matrices containing all the results for the supplementary individuals (coordinates, square cosine)

quanti.sup 	 a matrix containing the coordinates of the supplementary quantitative variables (the correlation between a variable and an axis is equal to the variable coordinate on the axis)

quali.sup 	 a list of matrices with all the results for the supplementary qualitative variables (coordinates of each categories of each variables, square cosine and v.test which is a criterion with a Normal distribution)

call 	 a list with some statistics

Returns the individuals factor map and the variables factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

print.MCA, plot.MCA, dimdesc

Examples


## Not run: 
data (poison)
MCA (poison, quali.sup = 3:4, quanti.sup = 1:2)
## End(Not run)

[Package FactoMineR version 1.05 Index]