MFA(FactoMineR)	R Documentation

Multiple Factor Analysis (MFA)

Description

Performs Multiple Factor Analysis (MFA) with supplementary individuals and supplementary groups of variables. Groups of variables can be quantitative or qualitative.

Usage


MFA (base, group, type = rep("s",length(group)), ind.sup = NULL, 
    ncp = 5, name.group = NULL, num.group.sup = NULL, 
    graph = TRUE, weight.col.mfa = NULL, axes = c(1,2))

Arguments

base 	 a data frame with n rows (individuals) and p columns (variables)

group 	 a list indicating the number of variables in each group

type 	 the type of variables in each group; three possibilities: "c" or "s" for quantitative variables (the difference is that for "s" variables are scaled to unit variance), "n" for qualitative variables; by default, all variables are quantitative and scaled to unit variance

ind.sup 	 a vector indicating the indexes of the supplementary individuals

ncp 	 number of dimensions kept in the results (by default 5)

name.group 	 a vector containing the name of the groups (by default, NULL and the group are named group.1, group.2 and so on)

num.group.sup 	 the indexes of the illustrative groups (by default, NULL and no group are illustrative)

graph 	 boolean, if TRUE a graph is displayed

weight.col.mfa 	 vector of weights, useful for HMFA method (by default, NULL and an MFA is performed)

axes 	 a length 2 vector specifying the components to plot

Value

summary.quali 	 a summary of the results for the qualitative variables

summary.quanti 	 a summary of the results for the quantitative variables

separate.analyses 	 the results for the separate analyses

eig 	 a numeric vector containing all the eigenvalues

group 	 a list of matrices containing all the results for the groups (Lg and RV coefficients, coordinates, square cosine, contributions, distance to the origin, the correlations between each group and each factor)

rapport.inertie 	 inertia ratio

ind 	 a list of matrices containing all the results for the active individuals (coordinates, square cosine, contributions)

ind.sup 	 a list of matrices containing all the results for the supplementary individuals (coordinates, square cosine)

quanti.var 	 a list of matrices containing all the results for the quantitative variables (coordinates, correlation between variables and axes)

quali.var 	 a list of matrices containing all the results for the supplementary qualitative variables (coordinates of each categories of each variables, and v.test which is a criterion with a Normal distribution)

partial.axes 	 a list of matrices containing all the results for the partial axes (coordinates, correlation between variables and axes, correlation between partial axes)

Returns the individuals factor map, the variables factor map and the groups factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

References

Escofier, B. and Pagès, J. (1994) Multiple Factor Analysis (AFMULT package), Computational Statistics and Data Analysis, 18, 121-140.

See Also

print.MFA, plot.MFA, dimdesc

Examples


data(wine)
aa = MFA(wine, group=c(2,5,3,10,9,2), type=c("n",rep("s",5)),
    ncp=5, name.group=c("orig","olf","vis","olfag","gust","ens"),
    num.group.sup=c(1,6))

## Not run: 
#### Interactive graph
liste = plot.MFApartial(aa)
plot(aa,choix="ind",habillage = "Terroir")

###Example 2
data (poison)
MFA(poison, group=c(2,2,5,6), type=c("s","n","n","n"),
    name.group=c("desc","desc2","symptom","eat"),
    num.group.sup=1:2)
## End(Not run)

[Package FactoMineR version 1.05 Index]