PCA(FactoMineR)	R Documentation

Principal Component Analysis (PCA)

Description

Performs Principal Component Analysis (PCA) with supplementary individuals, supplementary quantitative variables and supplementary qualitative variables.

Usage


PCA(X, scale.unit = TRUE, ncp = 5, ind.sup = NULL, 
    quanti.sup = NULL, quali.sup = NULL, row.w = NULL, 
    col.w = NULL, graph = TRUE, axes = c(1,2))

Arguments

X 	 a data frame with n rows (individuals) and p columns (numeric variables)

ncp 	 number of dimensions kept in the results (by default 5)

scale.unit 	 a boolean, if TRUE (value set by default) then data are scaled to unit variance

ind.sup 	 a vector indicating the indexes of the supplementary individuals

quanti.sup 	 a vector indicating the indexes of the quantitative supplementary variables

quali.sup 	 a vector indicating the indexes of the qualitative supplementary variables

row.w 	 an optional row weights (by default, uniform row weights)

col.w 	 an optional column weights (by default, uniform column weights)

graph 	 boolean, if TRUE a graph is displayed

axes 	 a length 2 vector specifying the components to plot

Value

Returns a list including:

eig 	 a numeric vector containing all the eigenvalues

var 	 a list of matrices containing all the results for the active variables (coordinates, correlation between variables and axes, square cosine, contributions)

ind 	 a list of matrices containing all the results for the active individuals (coordinates, square cosine, contributions)

ind.sup 	 a list of matrices containing all the results for the supplementary individuals (coordinates, square cosine)

quanti.sup 	 a list of matrices containing all the results for the supplementary quantitative variables (coordinates, correlation between variables and axes)

quali.sup 	 a list of matrices containing all the results for the supplementary qualitative variables (coordinates of each categories of each variables, and v.test which is a criterion with a Normal distribution)

Returns the individuals factor map and the variables factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

print.PCA, plot.PCA, dimdesc

Examples


data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup=13)
plot(res.pca,choix="ind",habillage=13)
dimdesc(res.pca, axes = 1:2)

[Package FactoMineR version 1.03 Index]