catdes(FactoMineR)	R Documentation

Categories description

Description

Description of the categories of one factor by qualitative variables and/or by quantitative variables

Usage


catdes(donnee,num.var,proba = 0.05)

Arguments

donnee 	 a data frame made up of at least one qualitative variables and a set of quantitative variables (sensory descriptors)

num.var 	 the indice of the variable to caracterized

proba 	 the significance threshold considered to caracterized the category (by default 0.05)

Value

Returns a list including:

test.chi 	 The qualitative variables which characterized the factor are listed in ascending order (from the one which characterized the most the factor to the one which significantly characterized with the proba proba

category 	 description of the each category of the num.var by each category of all the qualitative variables

quanti 	 the description of the each category of the num.var variable by the quantitative variables.

Author(s)

François Husson François.Husson@agrocampus-rennes.fr

References

Lebart, L., Morineau, A. and Piron, M. (1995) Statistique exploratoire multidimensionnelle, Dunod.

See Also

Examples


data(wine)
catdes(wine, num.var=2)

[Package FactoMineR version 1.03 Index]