plot.CA(FactoMineR)	R Documentation

Draw the Correspondance Analysis (CA) graphs

Description

Draw the Correspondance Analysis (CA) graphs.

Usage


plot.CA(x, axes = c(1, 2),
    xlim = NULL, ylim = NULL, invisible = NULL, col.row = "blue",
    col.col = "red", col.row.sup = "darkblue", 
    col.col.sup = "darkred", label = "all", cex = 1, 
    title = NULL, ...)

Arguments

x 	 an object of class CA

axes 	 a length 2 vector specifying the components to plot

xlim 	 range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim 	 range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

invisible 	 string indicating if some points should be unlabelled ("row", "col", "row.sup", "col.sup")

col.row 	 a color for the rows points

col.col 	 a color for columns points

col.row.sup 	 a color for the supplementary rows points

col.col.sup 	 a color for supplementary columns points

label 	 a list of character for the elements which are labelled (by default, all the elements are labelled ("row", "row.sup", "col", "col.sup"

cex 	 cf. function par in the graphics package

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

... 	 further arguments passed to or from other methods

Value

Returns the individuals factor map and the variables factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

CA

Examples


data(children)
res.ca <- CA (children, col.sup = 6:8, row.sup = 15:18)

[Package FactoMineR version 1.03 Index]