plot.DMFA(FactoMineR)	R Documentation

Make the Duale Multiple Factor Analysis (DMFA) graphs

Description

Plot the graphs for a Principal Component Analysis (DMFA) with supplementary individuals, supplementary quantitative variables and supplementary qualitative variables.

Usage


plot.DMFA(x, axes = c(1, 2), choix = "ind", label="all",
    invisible = NULL, ...)

Arguments

x 	 an object of class DMFA

axes 	 a length 2 vector specifying the components to plot

choix 	 the graph to plot ("ind" for the individuals, "var" for the variables)

label 	 a list of character for the elements which are labelled (by default, all the elements are labelled ("ind", ind.sup", "quali", "var", "quanti.sup"

invisible 	 string indicating if some points should not be drawn ("ind" or "quali")

... 	 further arguments passed to or from other methods

Value

Returns the individuals factor map and the variables factor map, the partial variables representation and the groups factor map.

Author(s)

Sébastien Lê, François Husson François.Husson@agrocampus-rennes.fr

See Also

DMFA

Examples



[Package FactoMineR version 1.02 Index]