plot.GPA(FactoMineR)	R Documentation

Draw the General Procrustes Analysis (GPA) map

Description

Draw the General Procrustes Analysis (GPA) map.

Usage


plot.GPA(x, axes = c(1, 2), 
    lab.ind.moy = TRUE, habillage = "ind",
    partial = "all", chrono = FALSE,
    xlim = NULL, ylim = NULL, cex = 1, title = NULL, ...)

Arguments

x 	 an object of class GPA

axes 	 a length 2 vector specifying the components to plot

lab.ind.moy 	 boolean, if TRUE, the label of the mean points are drwan

habillage 	 string corresponding to the color which are used. If "ind", one color is used for each individual; if "group" the individuals are colored according to the group

partial 	 list of the individuals or of the center of gravity for which the partial points should be drawn (by default, partial = "none" and no partial points are drawn)

chrono 	 boolean, if TRUE, the partial points of a same point are linked (useful when groups correspond to different moment)

xlim 	 range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim 	 range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

cex 	 cf. function par in the graphics package

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

... 	 further arguments passed to or from other methods

Value

Returns the General Procrustes Analysis map.

Author(s)

Elisabeth Morand, François Husson François.Husson@agrocampus-rennes.fr

See Also

GPA

[Package FactoMineR version 1.03 Index]