plot.MCA(FactoMineR)	R Documentation

Draw the Multiple Correspondance Analysis (MCA) graphs

Description

Draw the Multiple Correspondance Analysis (MCA) graphs.

Usage


plot.MCA (x, axes = c(1, 2), 
    xlim = NULL, ylim = NULL, invisible = NULL,
    col.ind = "blue", col.var = "red", col.quali.sup = "darkred",
    col.ind.sup = "darkblue", col.quanti.sup = "black",
    label = "all", cex = 1, title = NULL, habillage = "none", ...)

Arguments

x 	 an object of class MCA

axes 	 a length 2 vector specifying the components to plot

xlim 	 range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim 	 range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

invisible 	 string indicating if some points should not be drawn ("ind", "var", "ind.sup", "quali.sup", "quanti.sup")

col.ind 	 a color for the individuals, if color ="none" the label is not written

col.var 	 a color for the categories of qualitative variables, if color ="none" the label is not written

col.quali.sup 	 a color for the qualitative supplementary variables, if color ="none" the label is not written

col.ind.sup 	 a color for the supplementary individuals only if there is not habillage, if color ="none" the label is not written

col.quanti.sup 	 a color for the supplementary quantitative variables, if color ="none" the label is not written

label 	 print the labels of the points

cex 	 cf. function par in the graphics package

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

habillage 	 string corresponding to the color which are used. If "none", one color is used for the individual, another one for the qualitative variables; if "quali", one color is used for each qualitative variables; else if it is the position of a qualitative variable, it colors according to the different categories of this variable

... 	 further arguments passed to or from other methods

Value

Returns the individuals factor map and the variables factor map.

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

MCA

[Package FactoMineR version 1.04 Index]