plot.MFApartial(FactoMineR)	R Documentation

Plot an interactive Multiple Factor Analysis (MFA) graph

Description

Draw an interactive Multiple Factor Analysis (MFA) graphs.

Usage


plot.MFApartial(x, axes = c(1, 2), 
    lab.ind.moy = TRUE, lab.par = FALSE, habillage = "ind",
    chrono = FALSE, col.hab = NULL, invisible = NULL, 
    draw.partial = NULL, xlim = NULL, ylim = NULL, 
    cex = 1, title = NULL, ...)

Arguments

x 	 an object of class MFA

axes 	 a length 2 vector specifying the components to plot

lab.ind.moy 	 boolean, if TRUE, the label of the mean points are drwan

lab.par 	 boolean, if TRUE, the label of the partial points are drwan

habillage 	 string corresponding to the color which are used. If "ind", one color is used for each individual; if "quali" the individuals are colored according to one qualitative variable; if "group" the individuals are colored according to the group

chrono 	 boolean, if TRUE, the partial points of a same point are linked (useful when groups correspond to different moment)

col.hab 	 the colors to use. By default, colors are chosen

invisible 	 list of string; for choix ="ind", the individuals can be omit (invisible = "ind"), or supplementary individuals (invisible="ind.sup") or the centerg of gravity of the qualitative variables (invisible= "quali"); if invisible = c("ind","ind.sup"), just the centers of gravity are drawn

draw.partial 	 data frame of a boolean variable for all the individuals and all the centers of gravity and with for which the partial points should be drawn (by default, NULL and no partial points are drawn)

xlim 	 range for the plotted 'x' values, defaulting to the range of the finite values of 'x'

ylim 	 range for the plotted 'y' values, defaulting to the range of the finite values of 'y'

cex 	 cf. function par in the graphics package

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

... 	 further arguments passed to or from other methods

Value

Draw a graph with the individuals and the centers of gravity. The graph is interactive and clicking on a point will draw the partial points, if you click on a point for which the partial points are yet drawn, the partial points are deleted. To stop the interactive plot, click on the title (or in the top of the graph)

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

MFA, plot.MFA

Examples


## Not run: 
data(wine)
aa = MFA(wine,group=c(2,5,3,10,9,2),type=c("n",rep("s",5)),ncp=5,
    name.group=c("orig","olf","vis","olfag","gust","ens"),
    num.group.sup=c(1,6),graph=FALSE)
liste = plot.MFApartial(aa)
## End(Not run)

[Package FactoMineR version 1.00 Index]