poison(FactoMineR)	R Documentation

Poison

Description

The data used here refer to a survey carried out on a sample of children of primary school who suffered from food poisoning. They were asked about their symptoms and about what they ate.

Usage

data(poison)

Format

A data frame with 55 rows and 15 columns.

Source

Examples


data(poison)
res.mca <- MCA(poison, quanti.sup = 1:2, quali.sup=c(3,4))

[Package FactoMineR version 1.00 Index]