prefpls(FactoMineR)	R Documentation

Scatter plot and additional variables with quality of representation contour lines

Description

This function is useful to interpret the usual graphs $(x,y)$ with additional quantitative variables.

Usage


prefpls(donnee, var1 = 1, var2 = 2, firstvar = 3, 
    lastvar = ncol(donnee), levels = c(0.2,0.4,0.6,0.7,0.8,0.9,1), 
    asp = 1, nbchar = max(nchar(colnames(donnee))), title = NULL,
    choix="var")

Arguments

donnee 	 a data frame made up of quantitative variables

var1 	 the position of the variable corresponding to the x-axis

var2 	 the position of the variable corresponding to the y-axis

firstvar 	 the position of the first endogenous variable

lastvar 	 the position of the last endogenous variable (by default the last column of donnee)

levels 	 a list of the levels displayed in the graph of variables

asp 	 aspect ratio for the graph of the individuals

nbchar 	 the number of characters used for the labels of the variables

title 	 string corresponding to the title of the graph you draw (by default NULL and a title is chosen)

choix 	 the graph to plot ("ind" for the individuals, "var" for the variables)

Details

This function is very useful when there is a strong correlation between two variables x and y

Value

A scatter plot of the invividuals
A graph with additional variables and the quality of representation contour lines.

Author(s)

François Husson

References

Husson, F. & Pagès, J. (2005). Scatter plot and additional variables. Journal of applied statistics

Examples


data(decathlon)
prefpls(decathlon[,c(11,12,1:10)])

[Package FactoMineR version 1.02 Index]