print.PCA(FactoMineR)	R Documentation

Print the Principal Component Analysis (PCA) results

Description

Print the Principal Component Analysis (PCA) results.

Usage


## S3 method for class 'PCA':
print(x, file = NULL, sep = ";", ...)

Arguments

x 	 an object of class PCA

file 	 A connection, or a character string naming the file to print to. If NULL (the default), the results are not printed in a file

sep 	 character string to insert between the objects to print (if the argument file is not NULL

... 	 further arguments passed to or from other methods

Author(s)

Jérémy Mazet, François Husson François.Husson@agrocampus-rennes.fr

See Also

PCA, write.infile

Examples


## Not run: 
data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup = 13)
print(res.pca, file="c:/essai.csv", sep = ";")
## End(Not run)

[Package FactoMineR version 1.00 Index]