simule(FactoMineR)	R Documentation

Simulate by bootstrap

Description

Simulate by bootstrap

Usage


simule(data, nb.simul)

Arguments

data 	 A data frame from which the rows are the original data from which the simualte data are calculated (by the average of a bootstrap sample. The columns corresponds to the variables for which the simulation should be done. The first column must be a factor allowing to group the rows. A bootstrap simulation is done for each level of this factor.

nb.simul 	 The number of simulations.

Details

The simulation is independently done for each level of the factor. The number of rows can be different for each levels.

Value

mean 	 Data.frame with all the levels of the factor variable, and for each variable, the mean of the original data.

simul 	 Data.frame with all the levels of the factor variable, and for each variable, the nb.simul bootstrap simulations.

simul.mean 	 Data.frame with all the levels of the factor variable, and for each variable, the mean of the simulated data.

Author(s)

Jérémy Mazet

[Package FactoMineR version 1.00 Index]