svd.triplet(FactoMineR)	R Documentation

Singular Value Decomposition of a Matrix

Description

Compute the singular-value decomposition of a rectangular matrix with weights for rows and columns.

Usage


svd.triplet(X, Pl=NULL, Pc=NULL)

Arguments

X 	 a data matrix

Pl 	 vector with the weights of each row (NULL by default and the weights are uniform)

Pc 	 vector with the weights of each column (NULL by default and the weights are uniform)

Value

d 	 a vector containing the singular values of 'x';

u

v

{a matrix whose columns contain the right singular vectors of 'x'.}

See Also

svd

[Package FactoMineR version 1.00 Index]