write.infile(FactoMineR)	R Documentation

Print in a file

Description

Print in a file.

Usage


write.infile(X, file, sep=";", append = FALSE)

Arguments

X 	 an object of class list, data.frame, matrix, ...

file 	 A connection, or a character string naming the file to print to

sep 	 character string to insert between the objects to print (if the argument file is not NULL)

append 	 logical. If TRUE output will be appended to file; otherwise, it will overwrite the contents of file.

Author(s)

François Husson François.Husson@agrocampus-rennes.fr

Examples


## Not run: 
data(decathlon)
res.pca <- PCA(decathlon, quanti.sup = 11:12, quali.sup = 13)
write.infile(res.pca, file="c:/essai.csv", sep = ";")
## End(Not run)

[Package FactoMineR version 1.00 Index]